refactor(proxy): dedupe DeepSeek control-token stripper and inline reasoning buffer check
Export a single stripDeepSeekControlTokens from reasoning.ts, replacing the
identical private copies in openai-routes.ts and tool-call-stream.ts. Fold
shouldBufferImplicitReasoningContent into the now-private
shouldBufferImplicitReasoning in chat-completions-stream.ts, its only caller.
